frontend: refactor secret store to be more DRY
